+++ src/components/Common/Loader/SpinnerLoader.tsx
@@ -1,0 +1,8 @@
+import { Spinner } from '@chakra-ui/react';
+import React from 'react';
+
+const SpinnerLoader = () => {
+  return <Spinner size="lg" css={{ margin: '16px auto !important' }} />;
+};
+
+export default SpinnerLoader;
